+++ src/vue/modules/create-asset-form/mixins/manage-asset-request.mixin.js
@@ -3,7 +3,7 @@ import { uploadDocuments } from '@/js/helpers/upload-documents'
 import { base } from '@tokend/js-sdk'
 
 import { api } from '@/api'
-import { config } from '../_config'
+import config from '@/config'
 
 import { CreateAssetRequest } from '../wrappers/create-asset-request'
 
@@ -18,7 +18,7 @@ export default {
   computed: {
     preIssuanceAssetSigner () {
       return this.advancedStepForm.isPreissuanceDisabled
-        ? config().NULL_ASSET_SIGNER
+        ? config.NULL_ASSET_SIGNER
         : this.advancedStepForm.preIssuanceAssetSigner
     },
 
@@ -37,7 +37,7 @@ export default {
         code: this.informationStepForm.code,
         assetType: String(this.informationStepForm.assetType.value),
         preissuedAssetSigner: this.preIssuanceAssetSigner,
-        trailingDigitsCount: config().DECIMAL_POINTS,
+        trailingDigitsCount: config.DECIMAL_POINTS,
         initialPreissuedAmount: this.initialPreissuedAmount,
         maxIssuanceAmount: this.informationStepForm.maxIssuanceAmount,
         policies: this.informationStepForm.policies,