+++ src/vue/modules/create-sale-form/index.vue
@@ -70,8 +70,6 @@ import { ErrorHandler } from '@/js/helpers/error-handler'
 import { mapGetters } from 'vuex'
 import { vuexTypes } from '@/vuex'
 
-import { initConfig } from './_config'
-
 const STEPS = {
   information: {
     number: 1,
@@ -103,10 +101,6 @@ export default {
   },
   mixins: [LoadAssetsMixin, ManageSaleRequestMixin],
   props: {
-    storageUrl: {
-      type: String,
-      required: true,
-    },
     requestId: {
       type: String,
       default: '',
@@ -139,8 +133,6 @@ export default {
   methods: {
     async init () {
       try {
-        initConfig(this.storageUrl)
-
         await this.loadAssets(this.accountId)
         await this.tryLoadRequest()