+++ src/vue/modules/documents/document-upload-form/index.vue
@@ -92,8 +92,6 @@ import { DOCUMENT_TYPES } from '@/js/const/document-types.const'
 
 import { documentContainer, required } from '@validators'
 
-import { initConfig } from './_config'
-
 import { Bus } from '@/js/helpers/event-bus'
 import { ErrorHandler } from '@/js/helpers/error-handler'
 
@@ -119,13 +117,6 @@ export default {
     FormMixin,
     CreateAccountMixin,
   ],
-
-  props: {
-    storageUrl: {
-      type: String,
-      required: true,
-    },
-  },
 
   data: _ => ({
     form: {
@@ -181,10 +172,6 @@ export default {
 
       return 0
     },
-  },
-
-  created () {
-    initConfig(this.storageUrl)
   },
 
   methods: {