+++ src/vue/pages/AssetExplorer.vue
@@ -3,7 +3,6 @@
     <template v-if="getModule().canRenderSubmodule(AssetExplorerModule)">
       <submodule-importer
         :submodule="getModule().getSubmodule(AssetExplorerModule)"
-        :storage-url="storageUrl"
         :is-account-unverified="isAccountUnverified"
         :is-account-general="isAccountGeneral"
         :is-account-us-verified="isAccountUsVerified"
@@ -18,8 +17,6 @@
 import SubmoduleImporter from '@/modules-arch/submodule-importer'
 import { AssetExplorerModule } from '@modules/assets/asset-explorer/module'
 
-import config from '../../config'
-
 import { mapGetters } from 'vuex'
 import { vuexTypes } from '@/vuex'
 
@@ -31,7 +28,6 @@ export default {
 
   data: _ => ({
     AssetExplorerModule,
-    storageUrl: config.FILE_STORAGE,
   }),
   computed: {
     ...mapGetters({