+++ src/vue/pages/TradeExchange/Trade.OffersRenderer.vue
@@ -67,10 +67,10 @@
         </template>
       </template>
       <submit-trade-offer-form
-        :is-buy="isBuy"
+        :is-buy="!isBuy"
         :asset-pair="assetPair"
         :offer="selectedOffer"
-        @close-drawer="closeDrawer"
+        @offer-submitted="closeDrawer"
       />
     </drawer>
   </div>